[eas-cli] Rename observe:sessions to observe:session and require a session ID
The session ID is now a required positional argument and the
list-distinct-sessions mode is gone — along with --platform,
--start/--end/--days, --app-version, --update-id, --event-name, and the
helpers that only supported listing (SessionSummary, sort comparator,
list fetcher / formatter, event-count tracking). The command now exposes
only --project-id, --json, and --non-interactive.

Since the sessionId filter uniquely identifies the session, the time
range is no longer needed. Make startTime and endTime optional on
FetchObserveEventsOptions and FetchCustomEventsOptions and omit them
from the filter when undefined; the session fetcher relies on this so
the request is purely sessionId-scoped (existing observe:metrics and
observe:events callers still pass them explicitly, so behaviour is
unchanged for those commands).
