+++ packages/eas-cli/src/observe/__tests__/formatSessions.test.ts
@@ -1,120 +1,5 @@
-import { SessionEventEntry, SessionMetadata, SessionSummary } from '../fetchSessions';
-import {
-  buildObserveSessionEventsJson,
-  buildObserveSessionEventsTable,
-  buildObserveSessionListJson,
-  buildObserveSessionListTable,
-} from '../formatSessions';
-
-function makeSession(overrides: Partial<SessionSummary> = {}): SessionSummary {
-  return {
-    sessionId: 'session-1',
-    firstSeenAt: '2025-01-15T10:00:00.000Z',
-    lastSeenAt: '2025-01-15T10:05:00.000Z',
-    appVersion: '1.0.0',
-    appBuildNumber: '42',
-    deviceOs: 'iOS',
-    deviceOsVersion: '17.0',
-    deviceModel: 'iPhone 15',
-    ...overrides,
-  };
-}
-
-describe(buildObserveSessionListTable, () => {
-  it('shows session rows with first-seen timestamp and app version', () => {
-    const output = buildObserveSessionListTable([makeSession()], { daysBack: 7 });
-
-    expect(output).toContain('Sessions for the last 7 days');
-    expect(output).toContain('session-1');
-    expect(output).toContain('1.0.0 (42)');
-  });
-
-  it('omits event-count, Last seen, and Device columns in the table view', () => {
-    const output = buildObserveSessionListTable([makeSession()]);
-
-    expect(output).not.toContain('Metric events');
-    expect(output).not.toContain('Log events');
-    expect(output).not.toContain('Last seen');
-    expect(output).not.toContain('Device');
-    expect(output).not.toContain('iPhone 15');
-  });
-
-  it('includes the event name in the header when eventName option is set', () => {
-    const output = buildObserveSessionListTable([makeSession()], {
-      daysBack: 7,
-      eventName: 'login_pressed',
-    });
-
-    expect(output).toContain('Sessions containing event "login_pressed" for the last 7 days');
-  });
-
-  it('shows a tailored empty-state message when eventName is set', () => {
-    const output = buildObserveSessionListTable([], {
-      daysBack: 7,
-      eventName: 'login_pressed',
-    });
-    expect(output).toContain('No sessions containing event "login_pressed" found');
-  });
-
-  it('renders separate per-platform sections grouped by deviceOs', () => {
-    const sessions = [
-      makeSession({ sessionId: 'android-1', deviceOs: 'Android' }),
-      makeSession({ sessionId: 'ios-1', deviceOs: 'iOS' }),
-      makeSession({ sessionId: 'ios-2', deviceOs: 'iOS' }),
-    ];
-    const output = buildObserveSessionListTable(sessions);
-
-    const androidIndex = output.indexOf('Android');
-    const iosIndex = output.indexOf('iOS');
-    expect(androidIndex).toBeGreaterThanOrEqual(0);
-    expect(iosIndex).toBeGreaterThan(androidIndex);
-    expect(output).toContain('android-1');
-    expect(output).toContain('ios-1');
-    expect(output).toContain('ios-2');
-  });
-
-  it('renders an empty-state message when no sessions are present', () => {
-    const output = buildObserveSessionListTable([], {
-      daysBack: 7,
-      scannedMetricEventCount: 10,
-      scannedLogEventCount: 20,
-    });
-    expect(output).toContain('No sessions found');
-    expect(output).toContain('Scanned 10 metric events and 20 log events.');
-  });
-
-  it('includes a "Derived from N metric / M log events" footer', () => {
-    const output = buildObserveSessionListTable([makeSession()], {
-      scannedMetricEventCount: 50,
-      scannedLogEventCount: 80,
-    });
-    expect(output).toContain('Derived from 50 metric events and 80 log events.');
-  });
-
-  it('always includes the 100-event cap note', () => {
-    const output = buildObserveSessionListTable([makeSession()]);
-    expect(output).toContain('scans the most recent 100 metric events and 100 log events');
-  });
-
-  it('includes the 100-event cap note on the empty-result path too', () => {
-    const output = buildObserveSessionListTable([]);
-    expect(output).toContain('No sessions found');
-    expect(output).toContain('scans the most recent 100 metric events and 100 log events');
-  });
-});
-
-describe(buildObserveSessionListJson, () => {
-  it('returns a structured payload with scan counts and truncation flag', () => {
-    const sessions = [makeSession()];
-    const result = buildObserveSessionListJson(sessions, 10, 20, false);
-    expect(result).toEqual({
-      sessions,
-      scannedMetricEventCount: 10,
-      scannedLogEventCount: 20,
-      isTruncated: false,
-    });
-  });
-});
+import { SessionEventEntry, SessionMetadata } from '../fetchSessions';
+import { buildObserveSessionEventsJson, buildObserveSessionEventsTable } from '../formatSessions';
 
 function makeMetricEntry(overrides: Partial<SessionEventEntry> = {}): SessionEventEntry {
   return {
