+++ packages/eas-cli/src/observe/formatSessions.ts
@@ -1,112 +1,9 @@
 import chalk from 'chalk';
 
-import { SessionEventEntry, SessionMetadata, SessionSummary } from './fetchSessions';
-import { buildTimeRangeDescription, formatLogTimestamp } from './formatUtils';
+import { SessionEventEntry, SessionMetadata } from './fetchSessions';
+import { formatLogTimestamp } from './formatUtils';
 import { getMetricDisplayName } from './metricNames';
 import renderTextTable from '../utils/renderTextTable';
-
-export interface BuildSessionListOptions {
-  daysBack?: number;
-  startTime?: string;
-  endTime?: string;
-  eventName?: string;
-  scannedMetricEventCount?: number;
-  scannedLogEventCount?: number;
-  isTruncated?: boolean;
-}
-
-export function buildObserveSessionListTable(
-  sessions: SessionSummary[],
-  options?: BuildSessionListOptions
-): string {
-  const lines: string[] = [];
-  if (options) {
-    const timeDesc = buildTimeRangeDescription(options);
-    const subject = options.eventName
-      ? `Sessions containing event "${options.eventName}"`
-      : 'Sessions';
-    lines.push(chalk.bold(`${subject} ${timeDesc}`.trim()), '');
-  }
-
-  if (sessions.length === 0) {
-    const emptyMessage = options?.eventName
-      ? `No sessions containing event "${options.eventName}" found in the selected time range.`
-      : 'No sessions found in the selected time range.';
-    lines.push(chalk.yellow(emptyMessage));
-    if (options?.scannedMetricEventCount != null || options?.scannedLogEventCount != null) {
-      lines.push(
-        '',
-        `Scanned ${options.scannedMetricEventCount ?? 0} metric events and ${
-          options.scannedLogEventCount ?? 0
-        } log events.`
-      );
-    }
-    lines.push(
-      chalk.yellow(
-        'Note: this command scans the most recent 100 metric events and 100 log events; older sessions in this window may not be listed.'
-      )
-    );
-    return lines.join('\n');
-  }
-
-  const headers = ['Session ID', 'First seen', 'App version'];
-
-  const byPlatform = new Map<string, SessionSummary[]>();
-  for (const s of sessions) {
-    if (!byPlatform.has(s.deviceOs)) {
-      byPlatform.set(s.deviceOs, []);
-    }
-    byPlatform.get(s.deviceOs)!.push(s);
-  }
-
-  for (const [platform, platformSessions] of byPlatform) {
-    lines.push(chalk.bold(platform), '');
-    const rows: string[][] = platformSessions.map(s => [
-      s.sessionId,
-      formatLogTimestamp(s.firstSeenAt),
-      `${s.appVersion} (${s.appBuildNumber})`,
-    ]);
-    lines.push(renderTextTable(headers, rows), '');
-  }
-
-  if (options?.scannedMetricEventCount != null || options?.scannedLogEventCount != null) {
-    lines.push(
-      '',
-      `Derived from ${options.scannedMetricEventCount ?? 0} metric events and ${
-        options.scannedLogEventCount ?? 0
-      } log events.`
-    );
-  }
-
-  lines.push(
-    chalk.yellow(
-      'Note: this command scans the most recent 100 metric events and 100 log events; older sessions in this window may not be listed.'
-    )
-  );
-
-  return lines.join('\n');
-}
-
-export interface ObserveSessionListJson {
-  sessions: SessionSummary[];
-  scannedMetricEventCount: number;
-  scannedLogEventCount: number;
-  isTruncated: boolean;
-}
-
-export function buildObserveSessionListJson(
-  sessions: SessionSummary[],
-  scannedMetricEventCount: number,
-  scannedLogEventCount: number,
-  isTruncated: boolean
-): ObserveSessionListJson {
-  return {
-    sessions,
-    scannedMetricEventCount,
-    scannedLogEventCount,
-    isTruncated,
-  };
-}
 
 export interface BuildSessionEventsOptions {
   metadata?: SessionMetadata | null;
